fix optime

+++ dto/response-all.go
@@ -6,5 +6,5 @@ type ResponseALL struct {
 	RAM      float32 `json:"ram"`
 	Upload   float32 `json:"traffic_upload"`
 	Download float32 `json:"traffic_download"`
-	Uptime   int     `json:"uptime"`
+	Uptime   float32 `json:"uptime"`
 }
