+++ dto/response-uptime.go
@@ -1,6 +1,6 @@
 package dto
 
 type ResponseUptime struct {
-	Name string `json:"name"`
-	Max  int    `json:"max"`
+	Name string  `json:"name"`
+	Max  float32 `json:"max"`
 }
